Updated for GD SDK for iOS version 2.1.4439
Updated contributor application for GD SDK for iOS version 2.1.4439 and
Xcode 7.2.
Set target deployment to iOS 8, which is the minimum supported.
Updated copyright notices on launch screens.
Switched off Bitcode.
Added new icon size.
Changed .dylib to .tbd.
Removed getFileStat usage, which is no longer required.

+++ foriOS/src/system/KeychainUtility.m
@@ -573,7 +573,8 @@ secAttrValue_SecAccessControlRef:(SecAccessControlRef)secAccessControlRef
     }
     
     NSMutableString *ret = [NSMutableString stringWithFormat:
-                            @"%ld bytes %@", size, allPrint ? @"\"" : @"(" ];
+                            @"%ld bytes %@", (unsigned long)size,
+                            allPrint ? @"\"" : @"(" ];
     
     for (int index=0; index < size; index++) {
         if ( (!allPrint) && index > 0) {
